DefaultHandler extended in MyHandler private class.
